# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Roommate Chore & Expense Roulette API",
    "version": "1.0.0",
    "description": "# 🎲 Roommate Chore & Expense Roulette API\n\nProduction-grade FastAPI backend featuring specialized Data Structures & Algorithms (DSA):\n- **Chore Rotation**: Circular Queue (Ring Buffer) with O(1) modulo turn shift and sarcastic alerts\n- **Expense Ledger**: Hash Tables with O(1) net balance computations and strict zero-sum conservation\n- **Debt Simplification**: Min-Cash-Flow Greedy Directed Graph Algorithm reducing O(N^2) debts to <= N - 1 vectors\n- **Activity Undo Stack**: LIFO Stack for idempotent database state reversal\n- **Identity & Keys**: RFC 9562 UUIDv7 monotonic primary keys and Ed25519 asymmetric cryptography",
    "contact": {
      "name": "Roulette Engineering Team",
      "email": "engineering@roommate-roulette.local",
      "url": "https://github.com/roommate-roulette"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.roommate-roulette.local",
      "description": "Production Gateway"
    }
  ],
...
  "paths": {
...
    "/api/v1/households/": {
      "post": {
        "tags": ["Households & Memberships"],
        "summary": "Create a household entity",
        "description": "Creates an apartment or communal living entity with an RFC 9562 UUIDv7 primary key.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "201": {
            "description": "Household created"
          }
        }
      }
    },
    "/api/v1/households/{household_id}": {
      "get": {
        "tags": ["Households & Memberships"],
        "summary": "Get household by UUID7",
        "description": "Retrieves household entity and registered members with their turn queue indices.",
        "parameters": [
          {
            "name": "household_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Household retrieved"
          },
          "404": {
            "description": "Household not found"
          }
        }
      }
    },
    "/api/v1/households/{household_id}/members": {
      "post": {
        "tags": ["Households & Memberships"],
        "summary": "Register roommate into circular queue",
        "description": "Registers a user into the household circular ring buffer, assigning their deterministic turn order index.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "household_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "201": {
            "description": "Member added to circular queue"
          }
        }
      }
    },
...
    "/api/v1/expenses/": {
      "post": {
        "tags": ["Expenses & Debt Simplification"],
        "summary": "Record shared expense & splits into Hash Table ledger",
        "description": "Records an expense, tabulates individual participant splits, and pushes a mutation frame onto the LIFO activity undo stack.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "201": {
            "description": "Expense logged successfully"
          },
          "400": {
            "description": "Invalid splits sum"
          }
        }
      }
    },
    "/api/v1/expenses/{household_id}/balances": {
      "get": {
        "tags": ["Expenses & Debt Simplification"],
        "summary": "Compute Hash Table net balance vector O(1)",
        "description": "Returns O(1) net balance dictionary where positive represents money owed to user and negative represents money user owes. Guarantees sum of balances = 0.",
        "parameters": [
          {
            "name": "household_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Net balances tabulated"
          }
        }
      }
    },
    "/api/v1/expenses/{household_id}/simplify": {
      "get": {
        "tags": ["Expenses & Debt Simplification"],
        "summary": "Execute Min-Cash-Flow Directed Graph Debt Simplification",
        "description": "Executes greedy bipartite matching algorithm to compress multi-party debt webs into at most N - 1 minimal settlement transactions.",
        "parameters": [
          {
            "name": "household_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Debt graph simplified"
          }
        }
      }
    },
    "/api/v1/expenses/{household_id}/undo": {
      "post": {
        "tags": ["Expenses & Debt Simplification"],
        "summary": "Pop and undo top mutation from Activity Stack (LIFO)",
        "description": "Pops top record from ActivityStackLog and idempotently reverses the database state.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "household_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Mutation successfully undone"
          },
          "400": {
            "description": "Stack Underflow error with Sarcastic Alert"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "PyJWT Bearer token"
      },
      "Ed25519Signature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Ed25519-Signature",
        "description": "64-byte Ed25519 signature over server challenge nonce"
      },
      "Ed25519PublicKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Ed25519-Public-Key",
        "description": "32-byte Ed25519 hex public key"
      }
    },
...
  }
}